A compiler is no longer necessary. - sysname is now the source of the system definition which can be a c++,   vf or so file. - there is an xml attribut to sysname in the .knut file that specifies the   file type, it is called 'type'. - if the 'type' attribute is set to type="vf0", the vector field file won't   be compiled, but expression will be evaluated runtime.

+++ knut/src/basecomp.h
@@ -3,6 +3,7 @@
 
 #include "constants.h"
 #include "mat4data.h"
+#include <iostream>
 
 class KNAbstractContinuation
 {
@@ -51,7 +52,7 @@ class KNCliContinuation : public KNAbstractContinuation
     virtual void storeCursor() { charsPrinted = 0; }
     virtual void clearLastLine()
     {
-      for (int k=0; k<charsPrinted; ++k) std::cout << "\b";
+      for (size_t k=0; k<charsPrinted; ++k) std::cout << "\b";
     }
     static void printException(const KNException& ex)
     {
@@ -68,7 +69,7 @@ class KNCliContinuation : public KNAbstractContinuation
     void dataUpdated() { }
   private:
     KNDataFile* output;
-    int         charsPrinted;
+    size_t charsPrinted;
 };
 
 #endif
